Dedup: pHash → mean per-channel RMS pixel diff (pageflip + meetcat)
pHash (mean+DCT 8x8 = 64-bit fingerprint) was solving the wrong
problem. It exists to find images that are *similar* — same photo
at different scales, brightness, mild colour shifts. Slide
deduplication is the opposite question: are these two captures
pixel-identical except for cursor drift and PNG compression noise?
The 64-bit DCT shadow let visually-distinct slides with shared
structural patterns (dark background, centered title, bullets in
the lower half) collide at small Hamming distances, producing
false-positive revisits.

RMS over the actual pixel buffer is the right primitive. Cursor
movement and compression noise contribute fractions of a unit on
the 0–255 scale; a fully different slide contributes tens of
units. The threshold reads directly as "mean per-channel pixel
difference."

**pageflip (last-frame gate)**:
- src/dedup.rs: replaced HasherConfig + ImageHash with a Vec<u8>
  buffer + rms_diff(). Threshold flag is now f64 (default 5.0)
  instead of u32 bits. Stale image_hasher dependency removed
  from Cargo.toml.
- SlideEvent.phash field stays in the schema (for now) but is
  always None on the wire.
- session_log.SlideDeduped's dist/threshold are f64.

**meetcat (revisit detector)**:
- revisits.go: rewritten. Decode the saved PNG, downsample
  (aspect-preserved) to ≤1 MP keeping full RGBA, store the
  buffer, RMS-compare against every prior on the next slide.
  Memory budget: ~3.6 MB per stored slide × ≲100 slides per
  meeting = ~360 MB upper bound, fine on a modern Mac.
- main.go: classify takes the slide PNG path now, returns
  (revisit, idx, dist, err). The non-revisit path logs the
  closest-prior RMS so the threshold can be tuned against real
  meetings.
- revisits_test.go: rewritten with synthetic PNG fixtures
  covering exact-match, distinct slides, structurally-similar-
  but-different (the user's reported false-positive shape),
  cursor jitter (must stay revisit), and missing files.

Companion target raised: 🎯T31 (video-detection auto-suspend).
With RMS dedup, a video clip embedded in a slide makes every
captured frame cross the threshold; the pipeline shouldn't
drown in five LLM round-trips per second on near-duplicates.
A rolling-window detector and a "video mode" suspend belongs
in pageflip; the target captures the design.

+++ src/session_log.rs
@@ -97,11 +97,14 @@ pub enum LogEvent {
         /// Stable within a session for correlation; opaque across sessions.
         path_hash: String,
     },
-    /// A frame was suppressed by the dedup gate.
+    /// A frame was suppressed by the dedup gate. `dist` is the mean
+    /// per-channel RMS difference (0–255 scale) between the new and
+    /// last-saved frames; `threshold` is the configured gate value
+    /// in the same units.
     SlideDeduped {
         t_ms: u64,
-        dist: u32,
-        threshold: u32,
+        dist: f64,
+        threshold: f64,
     },
     /// A batch of audio samples was dispatched to the transcriber.
     AudioBatch {
